improved RegExs for xsd:dateTime
added improved regular expressions to check for xsd:dateTime in the checks: 2.2.0. (4) Annotation Provenance, 2.2.0. (5) Annotation Provenance, 3.3.1. (1) Time State and 3.3.1. (2) Time State. It uses the regular expression from http://www.w3.org/TR/xmlschema11-2/#dateTime with an additional improvement: The day of the month is now not just limited to 01-31 in general but is dependent on the month itself. (However for February the 29 is always as valid date according to the RegEx, since modeling the leap year rules in RegExs is impossible/really difficult).
For the checks 2.2.0. (5) and 3.3.1. (2) the time zone fragment is not optional.
For the checks 2.2.0. (4) and 3.3.1. (1) the time is now required (as specified by xsd:dateTime)
All of the checks now allow for fractions of a seconds (this would previously cause a warning, despite it being valid xsd:dateTime)
Additionally updated the documentations of the elements. For the checks 2.2.0. (4) and 3.3.1. (1) the allowed syntax is shown. For  the checks 2.2.0. (5) and 3.3.1. (2) the syntax does not show the seconds in the time zone anymore (it was not checked for it, nor should it have done this).
I hope this helps. We only noticed the problem, because 3.3.1. (1) would complain about incorrect time zone setting, but it was actually the fractions of the seconds that it didn't allow.
http://www.xkcd.com/208/ (Sorry, I couldn't resist)